usePartyFooterData hook implemented, FssaiNumber and gst expiry date field added in add ledger form, shortcuts working in ledger, NumberInput common component is used reather than FormikInputField

+++ src/utilities/handleKeyDown.ts
@@ -1,10 +1,12 @@
 export const handleKeyDownCommon = (
   event: KeyboardEvent,
-  handleDelete: (row: any) => void,
+  handleDelete?: (row: any) => void,
   handleUpdate?: (row: any) => void,
   togglePopup?: (state: boolean) => void,
   selectedRow?: any,
-  setView?: (view: { type: string; data: any }) => void
+  setView?: (view: { type: string; data: any }) => void,
+  handleSave?: (data:any) => void,
+  dataToSave?: any,
 ) => {
   switch (event.key) {
     case 'Escape':
@@ -22,18 +24,26 @@ export const handleKeyDownCommon = (
       break;
     case 'd':
     case 'D':
-      if (event.ctrlKey && selectedRow) {
+      if (event.ctrlKey && selectedRow && handleDelete) {
         handleDelete(selectedRow);
       }
       break;
     case 'e':
     case 'E':
       if (event.ctrlKey && selectedRow) {
+        event.preventDefault();
         if (setView) {
           setView({ type: 'add', data: selectedRow });
         } else if (handleUpdate) {
           handleUpdate(selectedRow);
         }
+      }
+      break;
+    case's':
+    case 'S':
+      if (event.ctrlKey && dataToSave && handleSave) {
+        event.preventDefault();
+        handleSave(dataToSave);
       }
       break;
     default:
